refactor: merge runtime option types into runtime
