+++ src/bub/runtime.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from collections.abc import AsyncIterator
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from enum import StrEnum
 from typing import Any, Literal
 
@@ -39,6 +39,24 @@ class BubError(Exception):
         if self.details:
             payload["details"] = self.details
         return payload
+
+
+@dataclass(frozen=True)
+class RuntimeChoice:
+    """One selectable runtime value."""
+
+    id: str
+    name: str | None = None
+    description: str | None = None
+    meta: dict[str, Any] | None = None
+
+
+@dataclass(frozen=True)
+class RuntimeOptions:
+    """Runtime choices that a channel or adapter may present to a user."""
+
+    models: list[RuntimeChoice] = field(default_factory=list)
+    current_model: str | None = None
 
 
 @dataclass
